+++ src/rules/no-missing-eof-newline/__tests__/index.js
@@ -6,15 +6,11 @@ import rule, { ruleName, messages } from ".."
 const testRule = ruleTester(rule, ruleName)
 
 testRule(undefined, tr => {
+  tr.ok("")
   tr.ok("\n")
   tr.ok("a { color: pink; }\n")
   tr.ok("a { color: pink; }\n\n\n")
 
-  tr.notOk("", {
-    message: messages.rejected,
-    line: 1,
-    column: 1,
-  })
   tr.notOk("a { color: pink; }", {
     message: messages.rejected,
     line: 1,