+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Head
 
 * Fixed: v3 regression causing bug in `rule-properties-order` and potentially other rules that accept arrays as primary options.
+* Fixed: `no-missing-eof-newline` no longer complains about completely empty files.
 
 # 3.0.3